Add files via upload
These are the audio recordings of the drone. All of these recordings were made using Dr. Siderius' microphone array and the Ishmael software. The two file names beginning with "sound" are recordings made with the drone on a table top at a distance of 0.5m and a separation of 0.4m between the two microphones. 
File names beginning in 1, 2, 3, or 4 indicate a recording taken at the specified distance from the center of the microphone pair with the drone operating at its "hover" program.
File names beginning in 0, 45, 135, or 180 indicate a recording taken at 4m distance at the angle indicated measured from the plane of the two microphones (IE 0 degrees indicates the measurement was taken inline with the microphones from the left side). The drone was operating its "hover" program.
The file name "various" is a recording taken at 4m with the drone's throttle being manually increased/decreased.